Abstract test cases.
