+++ openfecwebapp/tests/selenium/base_test_class.py
@@ -2,7 +2,9 @@ import os
 import unittest
 
 from selenium import webdriver
+from selenium.webdriver.common.keys import Keys
 from selenium.common.exceptions import NoSuchElementException
+
 from nose.plugins.attrib import attr
 
 
@@ -65,3 +67,50 @@ class BaseTest(unittest.TestCase):
         except NoSuchElementException:
             return False
         return True
+
+
+class SearchPageTestCase(BaseTest):
+
+    def getFilterDivByName(self, name):
+        return self.driver.find_element_by_xpath('//*[@for="' + name + '"]/..')
+
+    def openFilters(self):
+        self.driver.find_element_by_id('filter-toggle').click()
+
+    def getColumn(self, index, data):
+        return [row.find_elements_by_tag_name('td')[index].text
+                for row in data.find_elements_by_tag_name('tr')]
+
+    def checkFilter(self, name, first_entry, second_entry,
+                    count, index, result):
+        self.driver.get(self.url)
+        self.openFilters()
+        div = self.getFilterDivByName(name)
+        # Hack: Scroll to expand button before clicking; otherwise expand fails
+        # in Chrome and breaks tests
+        self.driver.execute_script('''
+            $('body').animate({
+                scrollTop: $('[for="%s"]').offset().top
+            }, 0);
+        ''' % (name, ))
+        div.find_element_by_xpath('./div/a/div/b').click()
+        self.assertEqual(
+            div.find_element_by_xpath('./div').get_attribute('class'),
+            ('chosen-container chosen-container-single '
+             'chosen-with-drop chosen-container-active'))
+        div.find_element_by_tag_name('input').send_keys(first_entry)
+        self.assertEqual(
+            len(div.find_elements_by_tag_name('li')),
+            count)
+        div.find_element_by_tag_name('input').send_keys(second_entry)
+        div.find_element_by_tag_name('input').send_keys(Keys.ENTER)
+        self.assertEqual(
+            div.get_attribute('class'),
+            'field active')
+        self.driver.find_element_by_id('category-filters').submit()
+        results = (self.driver.find_elements_by_tag_name('tr'))
+        col = [y.find_elements_by_tag_name('td')[index]
+                .text for y in results[1:]]
+        self.assertEqual(
+            set(col),
+            {result})
